// src/forgetrustee/frontend/src/types/generated/shaclResultDTO.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ShaclResultDTO {
  focusNode?: string;
  path?: string;
  value?: string;
  sourceShape?: string;
  sourceConstraintComponent?: string;
  severity?: string;
  message?: string;
}